refactor(calculator): test package 구조 변경
